스토어 사진 sftp방식으로 변환

+++ mit/src/main/java/com/itbank/service/StoreService.java
@@ -1,6 +1,7 @@
 package com.itbank.service;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -18,6 +19,12 @@ import com.itbank.model.StoreDTO;
 import com.itbank.model.WishListDTO;
 import com.itbank.model.writingDAO;
 import com.itbank.model.writingDTO;
+import com.jcraft.jsch.Channel;
+import com.jcraft.jsch.ChannelSftp;
+import com.jcraft.jsch.JSch;
+import com.jcraft.jsch.JSchException;
+import com.jcraft.jsch.Session;
+import com.jcraft.jsch.SftpException;
 
 @Service
 public class StoreService {
@@ -25,72 +32,21 @@ public class StoreService {
    @Autowired private StoreDAO dao;
    @Autowired private writingDAO wdao;
    @Autowired private QnaDAO qnadao;
-   
-   private final String uploadPath = "C:\\mitImg";
-   
-   public StoreService() {
-      File dir = new File(uploadPath);
-      if(dir.exists() == false) {
-         dir.mkdirs();
-      }
-   }
+   @Autowired private ToSftpService tss;
    
 
-   public List<StoreDTO> getList() {
+   public List<StoreDTO> getList() throws JSchException, SftpException, IOException {
       List<StoreDTO> list = dao.getList();
      
+      for(StoreDTO dto : list) {
+    	  dto.setMainimg(tss.getimgToServer(dto.getMainimg()));
+      }
       return list;
    }
    
    // +================================+
 
    public int getInsert(StoreDTO dto) {
-      List<MultipartFile> files = dto.getFiles();
-      
-      if (files.get(0).getSize() == 0) {
-         dto.setPdimg("");
-         dto.setViewimg("");
-         return dao.getInsertItem(dto);
-      }
-      
-      // id이름에 따라서 폴더 생성
-      String newdir = uploadPath + "\\" + dto.getPdcode() + dto.getPdwriter();
-      String fileName = "";
-      String viewimgname = "";
-      List<String> viewimglist = dto.getViewimglist();
-      
-      
-      for (MultipartFile f : files) {
-         if(f.getSize() == 0) {
-            break;
-         }
-         UUID uuid = UUID.randomUUID();
-         String fileName2 = uuid.toString() + "_" + f.getOriginalFilename();
-         if(dto.getMainimg().equals(f.getOriginalFilename())) {
-        	 dto.setMainimg(fileName2);
-         }
-         for(String s : viewimglist) {
-        	 if(s.equals(f.getOriginalFilename())) {
-        		 viewimgname += fileName2 + ",";
-        		 break;
-        	 }
-         }
-         File dest = new File(newdir, fileName2);   // 파일 객체를 생성
-         if(dest.exists() == false) {
-             dest.mkdirs();
-         }
-         fileName += fileName2 + ",";
-         
-         try {
-            f.transferTo(dest);
-         } catch (IllegalStateException | IOException e) {
-            e.printStackTrace();
-         } 
-            
-      }
-      
-      dto.setViewimg(viewimgname);
-      dto.setPdimg(fileName);
       int row = dao.getInsertItem(dto);
       return row;
       
@@ -102,20 +58,19 @@ public class StoreService {
    }
 
 
-   public StoreDTO selectOne(int idx) {
+   public StoreDTO selectOne(int idx) throws JSchException, SftpException, IOException {
       StoreDTO dto = dao.SelecOne(idx);
       ArrayList<String> list2 = new ArrayList<String>();
       if(dto.getPdimg() != null){
-         File dir = new File(uploadPath + "\\" + dto.getPdcode() + dto.getPdwriter());
-         File[] files = dir.listFiles();
          String filelistname = dto.getPdimg();
+         dto.setMainimg(tss.getimgToServer(dto.getMainimg()));
          
-         for(int i = 0 ; i <  files.length ; i++) {
+         for(int i = 0 ; i <  filelistname.length() ; i++) {
         	 if(filelistname.length() <= 0 ) {
        		  break;
        	  	}
         	String msg = filelistname.substring(0, filelistname.indexOf(","));
-        	list2.add(msg);
+        	list2.add(tss.getimgToServer(msg));
         	filelistname = filelistname.substring(filelistname.indexOf(",")+1);        
          }
          dto.setFilename(list2);
@@ -124,16 +79,14 @@ public class StoreService {
       list2 = new ArrayList<String>();
       
       if(dto.getViewimg() != null){
-          File dir = new File(uploadPath + "\\" + dto.getPdcode() + dto.getPdwriter());
-          File[] files = dir.listFiles();
           String filelistname = dto.getViewimg();
           
-          for(int i = 0 ; i <  files.length ; i++) {
+          for(int i = 0 ; i <  filelistname.length() ; i++) {
         	  if(filelistname.length() <= 0 ) {
         		  break;
         	  }
         	 String msg = filelistname.substring(0, filelistname.indexOf(","));
-         	list2.add(msg);
+        	 list2.add(tss.getimgToServer(msg));
          	filelistname = filelistname.substring(filelistname.indexOf(",")+1);        
           }
           dto.setViewimglist(list2);
@@ -143,18 +96,8 @@ public class StoreService {
    }
 
 
-   public int delete(int idx) {
+   public int delete(int idx) throws JSchException, SftpException, IOException {
       StoreDTO dto = selectOne(idx);
-      
-      if(dto.getPdimg() != null) {
-    	  String newdir = uploadPath + "\\" + dto.getPdcode() + dto.getPdwriter();
-          File delfolder = new File(newdir);
-          File[] delfolderList = delfolder.listFiles();
-    	  for(int i = 0; i< delfolderList.length ; i++) {
-    	         delfolderList[i].delete();
-    	  }
-    	  delfolder.delete();
-      }
       
       int row = deleteallreply(dto.getIdx());
       
@@ -167,22 +110,16 @@ public class StoreService {
    }
 
 
-   private int deleteallreply(int idx) {
+   private int deleteallreply(int idx)  {
 	   List<writingDTO> list = wdao.getreplylist(idx);
+	   
+	   if(list == null) {
+		   return 1;
+	   }
 	   
 	   int row = 0;
 	   
 	   for(writingDTO dto : list) {
-		   
-		   if(dto.getImg() != null) {
-		    	  String newdir = uploadPath + "\\store" + dto.getWriter() + dto.getPdidx();
-		          File delfolder = new File(newdir);
-		          File[] delfolderList = delfolder.listFiles();
-		    	  for(int i = 0; i< delfolderList.length ; i++) {
-		    	         delfolderList[i].delete();
-		    	  }
-		    	  delfolder.delete();
-		      }
 		   row += wdao.deletereply(dto.getIdx());
 	   }
 	   return row;
@@ -200,39 +137,39 @@ public class StoreService {
    
 
 	public int insert(writingDTO dto) {
-		List<MultipartFile> files = dto.getWritingfiles();
-	      
-	    if (files == null) {
-	       dto.setImg("");
-	       return wdao.insert(dto);
-	    }
-	    
-	    String newdir = uploadPath + "\\store" + dto.getWriter() + dto.getPdidx();
-	    String fileName = "";
-	      
-	      for (MultipartFile f : files) {
-	         if(f.getSize() == 0) {
-	            break;
-	         }
-	         
-	        // 랜덤 파일명 
-			UUID uuid = UUID.randomUUID();
-			String fileName2 = uuid.toString() + "_" + f.getOriginalFilename();
-	       
-	         File dest = new File(newdir, fileName2);   // 파일 객체를 생성
-	         fileName += fileName2 + ",";
-	         
-	         if(dest.exists() == false) {
-	            dest.mkdirs();
-	         }
-	         try {
-	            f.transferTo(dest);
-	         } catch (IllegalStateException | IOException e) {
-	            e.printStackTrace();
-	         } 
-	            
-	      }
-	    dto.setImg(fileName);
+//		List<MultipartFile> files = dto.getWritingfiles();
+//	      
+//	    if (files == null) {
+//	       dto.setImg("");
+//	       return wdao.insert(dto);
+//	    }
+//	    
+//	    String newdir = uploadPath + "\\store" + dto.getWriter() + dto.getPdidx();
+//	    String fileName = "";
+//	      
+//	      for (MultipartFile f : files) {
+//	         if(f.getSize() == 0) {
+//	            break;
+//	         }
+//	         
+//	        // 랜덤 파일명 
+//			UUID uuid = UUID.randomUUID();
+//			String fileName2 = uuid.toString() + "_" + f.getOriginalFilename();
+//	       
+//	         File dest = new File(newdir, fileName2);   // 파일 객체를 생성
+//	         fileName += fileName2 + ",";
+//	         
+//	         if(dest.exists() == false) {
+//	            dest.mkdirs();
+//	         }
+//	         try {
+//	            f.transferTo(dest);
+//	         } catch (IllegalStateException | IOException e) {
+//	            e.printStackTrace();
+//	         } 
+//	            
+//	      }
+//	    dto.setImg(fileName);
 		return wdao.insert(dto);
 	}
 
@@ -303,25 +240,13 @@ public class StoreService {
 	}
 
 
-	public List<writingDTO> getreplylist(int idx) {
+	public List<writingDTO> getreplylist(int idx) throws JSchException, SftpException, IOException{
 		
 		List<writingDTO> list = wdao.getreplylist(idx);
-		ArrayList<String> list2 = new ArrayList<String>();
 		
 		for(writingDTO dto : list) {
 			if(dto.getImg() != null) {
-				File dir = new File(uploadPath + "\\store" + dto.getWriter() + + dto.getPdidx());
-		        File[] files = dir.listFiles();
-		        String filelistname = dto.getImg();
-		        for(int i = 0 ; i <  files.length ; i++) {
-		        	 if(filelistname.length() <= 0 ) {
-		       		  break;
-		       	}
-		        String msg = filelistname.substring(0, filelistname.indexOf(","));
-			    list2.add(msg);
-			    filelistname = filelistname.substring(filelistname.indexOf(",")+1);   
-			}
-		    dto.setWritingfilename(list2);
+			dto.setWritingfilename(tss.getimgToServer(dto.getImg()));
 			}
 		}
 		return list;
@@ -353,56 +278,11 @@ public class StoreService {
 
 
 	public int modifyreply(writingDTO dto) {
-		List<MultipartFile> files = dto.getWritingfiles();
-	      
-	    if (files == null) {
-	       dto.setImg("");
-	       return wdao.modifyreply(dto);
-	    }
-	    
-	    String newdir = uploadPath + "\\store" + dto.getWriter() + dto.getPdidx();
-	    String fileName = "";
-	      
-	      for (MultipartFile f : files) {
-	         if(f.getSize() == 0) {
-	            break;
-	         }
-	         
-	        // 랜덤 파일명 
-			UUID uuid = UUID.randomUUID();
-			String fileName2 = uuid.toString() + "_" + f.getOriginalFilename();
-	       
-	         File dest = new File(newdir, fileName2);   // 파일 객체를 생성
-	         fileName += fileName2 + ",";
-	         
-	         if(dest.exists() == false) {
-	            dest.mkdirs();
-	         }
-	         try {
-	            f.transferTo(dest);
-	         } catch (IllegalStateException | IOException e) {
-	            e.printStackTrace();
-	         } 
-	            
-	      }
-	    dto.setImg(fileName);
-		
 		return wdao.modifyimgreply(dto);
 	}
 
 
 	public int deletereply(int idx) {
-		writingDTO dto = wdao.getreplyone(idx);
-		
-		if(dto.getImg() != null) {
-	    	  String newdir = uploadPath + "\\store" + dto.getWriter() + dto.getPdidx();
-	          File delfolder = new File(newdir);
-	          File[] delfolderList = delfolder.listFiles();
-	    	  for(int i = 0; i< delfolderList.length ; i++) {
-	    	         delfolderList[i].delete();
-	    	  }
-	    	  delfolder.delete();
-	      }
 		return wdao.deletereply(idx);
 	}
